Fix: Handle 401 error on apply for card page
The `apply-for-card.js` script was making a request to `/api/users/me` on page load, which caused an unhandled 401 error for non-logged-in users.

This commit fixes the issue by adding a `.catch` block to the fetch request to handle the error gracefully. It also improves the role-checking logic to be more robust.
